requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
blinker==1.7.0
cachelib==0.13.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.2
Flask-Session==0.8.0
google-ai-generativelanguage==0.4.0
google-api-core==2.16.2
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.1
grpcio-status==1.60.1
idna==3.6
install==1.3.5
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.1.21
langchain-google-genai==0.0.7
langsmith==0.0.87
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
msgspec==0.18.6
mysql-connector-python==8.3.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.29
stack-data==0.6.3
tenacity==8.2.3
threadpoolctl==3.4.0
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
